# openapi.yaml
openapi: 3.0.3
info:
  title: WEBonTour API
  description: API for the WEBonTour CMS
  version: 1.0.0
  contact:
    name: Project Team
  license:
    name: Project License
servers:
  - url: /api
    description: API base path
...
paths:
...
  /posts:
    get:
      tags:
        - posts
      summary: Get all posts
      description: Get a list of all posts with optional filtering
      operationId: getPosts
      parameters:
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category name
        - name: tag
          in: query
          schema:
            type: string
          description: Filter by tag name
        - name: author
          in: query
          schema:
            type: string
          description: Filter by author username
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Number of posts to return
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of posts to skip
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [date]
            default: date
          description: Sort by field
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order
      responses:
        '200':
          description: Posts retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  posts:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
                  pagination:
                    type: object
                    properties:
                      currentPage:
                        type: integer
                        example: 1
                      totalPages:
                        type: integer
                        example: 5
                      totalPosts:
                        type: integer
                        example: 42
                      hasNextPage:
                        type: boolean
                        example: true
                      hasPrevPage:
                        type: boolean
                        example: false
                      limit:
                        type: integer
                        example: 20
                  sorting:
                    type: object
                    properties:
                      sortBy:
                        type: string
                        example: date
                      sortOrder:
                        type: string
                        example: desc
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - posts
      summary: Create a new post
      description: Create a new blog post (writer or admin)
      operationId: createPost
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - title
                - content
                - category_id
              properties:
                title:
                  type: string
                  example: My New Blog Post
                content:
                  type: string
                  example: This is the content of my blog post
                category_id:
                  type: integer
                  example: 1
                tags:
                  type: array
                  items:
                    type: integer
                  example: [1, 2, 3]
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: Attachments for the post
                custom_date:
                  type: string
                  format: date-time
                  description: Custom publication date (admin only)
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Bad request - validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized (writer or admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /posts/{id}:
    get:
      tags:
        - posts
      summary: Get post by ID
      description: Get a specific post by its ID
      operationId: getPostById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Post ID
      responses:
        '200':
          description: Post retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - posts
      summary: Update a post
      description: Update an existing blog post (writer or admin)
      operationId: updatePost
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Post ID
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Updated Blog Post Title
                content:
                  type: string
                  example: This is the updated content
                category_id:
                  type: integer
                  example: 2
                tags:
                  type: array
                  items:
                    type: integer
                  example: [2, 3, 4]
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                  description: New attachments for the post
                removeAttachments:
                  type: array
                  items:
                    type: integer
                  description: IDs of attachments to remove
                custom_date:
                  type: string
                  format: date-time
                  description: Custom publication date (admin only)
      responses:
        '200':
          description: Post updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Bad request - validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized (writer or admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - posts
      summary: Delete a post
      description: Delete an existing blog post (writer or admin)
      operationId: deletePost
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Post ID
      responses:
        '200':
          description: Post deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Post deleted successfully
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized (writer or admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Post not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /comments:
    post:
      tags:
        - comments
      summary: Create a comment
      description: Create a new comment on a post
      operationId: createComment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - postId
                - content
              properties:
                postId:
                  type: integer
                  example: 1
                content:
                  type: string
                  maxLength: 2000
                  example: This is a great post!
                parentId:
                  type: integer
                  example: 5
                  description: ID of parent comment for replies
      responses:
        '201':
          description: Comment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Bad request - validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Post or parent comment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /comments/{id}:
    put:
      tags:
        - comments
      summary: Update a comment
      description: Update an existing comment (author only or admin/moderator)
      operationId: updateComment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Comment ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  maxLength: 2000
                  example: Updated comment content
      responses:
        '200':
          description: Comment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Bad request - validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized to edit this comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - comments
      summary: Delete a comment
      description: Delete a comment (soft delete - author only or admin/moderator)
      operationId: deleteComment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Comment ID
      responses:
        '200':
          description: Comment deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Comment deleted successfully
        '400':
          description: Comment already deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized to delete this comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Comment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /signup-keys:
    get:
      tags:
        - signup-keys
      summary: Get all signup keys
      description: Get a list of all signup keys (admin only)
      operationId: getSignupKeys
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Signup keys retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SignupKey'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized (admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - signup-keys
      summary: Create a new signup key
      description: Create a new signup key (admin only)
      operationId: createSignupKey
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                note:
                  type: string
                  example: For John from HR
                  description: Optional note to track the key purpose
      responses:
        '201':
          description: Signup key created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignupKey'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized (admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /signup-keys/{id}:
    put:
      tags:
        - signup-keys
      summary: Update signup key note
      description: Update the note for a signup key (admin only)
      operationId: updateSignupKey
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Signup key ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                note:
                  type: string
                  example: Updated note
      responses:
        '200':
          description: Signup key updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignupKey'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized (admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Signup key not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - signup-keys
      summary: Delete signup key
      description: Delete a signup key (admin only)
      operationId: deleteSignupKey
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Signup key ID
      responses:
        '200':
          description: Signup key deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Signup key deleted successfully
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized (admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Signup key not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /site-settings:
    get:
      tags:
        - site-settings
      summary: Get all site settings
      description: Get all site settings (admin only)
      operationId: getSiteSettings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Site settings retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SiteSetting'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized (admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /site-settings/{key}:
    put:
      tags:
        - site-settings
      summary: Update site setting
      description: Update a specific site setting (admin only)
      operationId: updateSiteSetting
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
          description: Setting key
          example: site_visibility
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - value
              properties:
                value:
                  type: string
                  example: public
      responses:
        '200':
          description: Site setting updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteSetting'
        '400':
          description: Bad request - invalid value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not authorized (admin only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Setting not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Admin endpoints
...
  /document/{filename}:
    get:
      tags:
        - documents
      summary: Get document
      description: Get a legal document or policy by filename
      operationId: getDocument
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
          description: Document filename without extension
          example: privacy
      responses:
        '200':
          description: Document content
          content:
            text/markdown:
              schema:
                type: string
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Post:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: My Blog Post
        content:
          type: string
          example: This is the content of my blog post...
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        author:
          type: string
          example: johndoe
        author_id:
          type: integer
          example: 1
        category_id:
          type: integer
          example: 1
        category_name:
          type: string
          example: Technology
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
...
    Tag:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: JavaScript
...
    Comment:
      type: object
      properties:
        id:
          type: integer
          example: 1
        content:
          type: string
          example: This is a great post!
        parent_id:
          type: integer
          example: null
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        is_deleted:
          type: boolean
          example: false
        author:
          type: string
          example: johndoe
        author_id:
          type: integer
          example: 1
        replies:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
    SignupKey:
      type: object
      properties:
        id:
          type: integer
          example: 1
        key_value:
          type: string
          example: abc123def456ghi789...
        note:
          type: string
          example: For John from HR
          nullable: true
        created_at:
          type: string
          format: date-time
        created_by_username:
          type: string
          example: admin
...
    Error:
      type: object
      properties:
        message:
          type: string
          example: Error message